Update patch set 1
Patch Set 1:

It appears VID-38 changed the code to conform with the typo so fixing the typo will now require VID to be updated. I had issues pushing this code earlier so the issue was patched between me fixing the issue and getting the code pushed. If there is a desire to fix the typo this can be put in and VID updated, not sure if other code is referencing the typo as well though.

Patch-set: 1
Reviewer: Gerrit User 1571 <1571@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0