Update patch set 2
Patch Set 2: Code-Review-2

It cannot use a log framework to write warnings, because IT IS a logging framework. The only way to communicate with a user is System.out/System.err. Please leave it as is for now, let's fix the Sonar violation later if needed.

Patch-set: 2
Reviewer: Gerrit User 329 <329@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2