Update patch set 1
Patch Set 1: Code-Review-2

(36 comments)

this is impossible to review. I've got tiered and stopped at class UnifiedCompositionService.java

there are a lot of comments to fix.
fix all of them and I'll continue the review on Sunday.
I suggest that you'll take my comments and check if this issues exist in the classes I did not checked yet and fix them

Patch-set: 1
Label: Code-Review=-2