Update patch set 1
Patch Set 1:

We don't want to delete the node_modules folder each time a local build is running because it doubles the time it takes for the project to build.

This was commented out already and was put back into the code for some reason.

Patch-set: 1
Reviewer: Gerrit User 1341 <1341@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0