Update patch set 8
Patch Set 8:

Hi, Guys
It seems that the discussion of whether "validationVsp is static" or not is out of scope of this patch. I consider that we can open another JIRA/patch to discuss and track this "validationVsp is static" issue. if the final conclusion is "validationVsp shall not be static", then this patch will not be valid any more, I will close it.

Patch-set: 8