Update patch set 2
Patch Set 2:

(2 comments)

please see my comments, i need to delay the merging of this since we are in the process of rearranging the dockers on our end. we are separating the catalog from the onboarding once that is merged you will need to rebase and update the template location.

Patch-set: 2
Label: Code-Review=0