Update patch set 1
Patch Set 1:

i do not thsink you should do it one by one. the node types are depandent on the capabelties and data types. create one commit with all the change and work on it and it will be aligne and ervery one will be able to look ate it.

Patch-set: 1