Update patch set 1
Patch Set 1:

(10 comments)

I think all common plugins should be declared in pluginManagement. This is much cleaner, will ease on the child POMs, and get rid of all those properties. With the additional benefit of enforcing the same version and thus consistent behavior throughout the project. Also, a groupId be must be included in every reference to a plugin, even if it was declared in pluginManagement. See https://maven.apache.org/pom.html#Plugin_Management

Patch-set: 1
Label: Verified=0