Update patch set 1
Patch Set 1:

Hi Orit. Upgrading everything to ONAP-UI would be a lot of additional work unrelated to the purpose of this commit. I understand it needs to be done, but with this commit we are just trying to add a feature to the existent operation screen. Let's create a JIRA task and do the enhancement in a separate commit.

Patch-set: 1
Reviewer: Gerrit User 3411 <3411@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0