Update patch set 2
Patch Set 2:

This code change is to reduce the code that is considered for the coverage, as its clear this is not the production code to be considered for the overall coverage.
I will have another commit where I will include the test code from test folder.

Patch-set: 2
Label: Code-Review=0