Update patch set 4
Patch Set 4:

Hi Michael,

The issue here arrives even before certifying the VF. If we try to change the name of VF before certifying than it doesn't work if there are any operations created in side it. There is a bug in VF operation handling which causes this.

Thanks
Priyanshu

Patch-set: 4