Update patch set 4
Patch Set 4:

you are working on a mac where your source code and dockers run on the same machine for most of the devalopers they have avirtual box for the dockers and the source is out side of it.

i do not belive the logic for the auto json needs to be auto generated.

Patch-set: 4