Update patch set 7
Patch Set 7:

i do not believe this is the correct way to go since you do not handle a case where we are building from different branch in addition the configuration files will be overridden each time the script runs and will not allow some to pull from a different repo. the manual step. i do not feel this adds any value and creates more issues then it solves.
We have been working with the script for awhile and this is not the issue we see with it.

Patch-set: 7