Update patch set 6
Patch Set 6:

i understand but i do not believe in over loading our log by default with all this info in case of an error by default.
In case of debugging a user can simply change the log level.
There were points where your change is welcomed and I marked them the other places please leave it as is.
As a general step for us in sdc if we want to debug we simply change the log level to debug.

Patch-set: 6