Update patch set 6
Patch Set 6:

changing the log level to debug would be the last thing I will do, it's definitely a nightmare especially for a new developer if you ever try it. That means developers have to search some logs they could be interested in in a huge number of log lines like finding a needle in a bundle of hay. For a sdc veteran like you, it may be easy as a pie, but for newcomers, it is not a friendly way to debug issues.

Patch-set: 6
Label: Code-Review=0