Update patch set 9
Patch Set 9:

> (6 comments)

Line 598:
sdc already have a modal for deleting component instances from the graph.
it is only for single node,same modal instance is used as here it is multiple nodes

Line 614:
Please add an copiedInstance property to the scope instead of define everything here.
This is required by backend in this format ,as comment is given by Michel to send in this format from frontend

Patch-set: 9