Update patch set 10
Patch Set 10:

if a developer changes the debug log level to DEBUG, that would be a disaster which means a developer have to find a log he/she wants from tens of thousands of log lines. So this way does not make sense at all at least for a new developer.

Patch-set: 10