Update patch set 11
Patch Set 11:

Hi,

I have a question regarding the design.
In a scenario in which I copy an instance and then not immediately pasting it instead I continue working in the composition tab and after a while I want to paste it, or even the opposite. I accidentally paste it, I don't have any visual sign of which instance has been copied.
In more complex situation this could be a problem for the user.
Is this something that is resolvable? 
Adding some kind of visual representation of the copied instance?

Patch-set: 11
Reviewer: Gerrit User 1341 <1341@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0