Update patch set 12
Patch Set 12:

Hi, please apply the following:

Position of the pasted item-
After pasting, the pasted item should be placed in a constant place.
We may want to place the pasted item right next to the copied item so even when there are many items on the canvas, the user would know where to find it.

Selected item after pasting-
Current behavior: After pasting, the original item is selected
Desired behavior: after pasting, the pasted item is selected.

Thanks

Patch-set: 12