Update patch set 2
Patch Set 2:

Yuriy, if the artifact is only relevany for VF/PNF and not VL CP etc I'd rather to keep it like it is now.

I also wanted to make sure you understand the difference between resource and resource instance: do we need CDS artifacts in both cases?

Patch-set: 2