Update patch set 2
Patch Set 2:

Resource is any VF/PNF as a building block, once dragged & dropped to a service composition, it becomes a resource instance.

Any definitions (hence properties, artifacts) that are part of the resource are automatically inherited to the resource instance level, but not the other way around.

Question is, should a CDS artifact be generated per a VF/PNF, or maybe it might change within instances of the same VF/PNF inside different services?

Patch-set: 2