Update patch set 2
Patch Set 2:

Excellent question, i could see the need for both use case. The 80%/20% rule will be that the CBA artifact is part of the model that is reusable across all service for VF/PNF. However, the 20% of use case that might require a customization of the CBA Artifact for a given VF/PNF within a service.

Patch-set: 2