Supporting PNFD (SOL001) mapping to Internal AID model
This maps the PNF descriptor containing etsi defined types
according to SOL001 v2.5.1 to openecomp internal model. The
transformation is done in the "Create VSP" stage. As the main
PNF type gets mapped to the outer container in internal ecomp
model, it is possible to have no "resource instances" when
importing the PNF model into the catalog. The code was updated
to allow this for PNFs.

Change-Id: I961e91746ca638d858dd468e6c82aeb57993f7a3
Issue-ID: SDC-1976
Issue-ID: SDC-2160
Signed-off-by: dermot123 <dermot.melia@est.tech>