Update patch set 2
Patch Set 2:

(4 comments)

> please go over the SoftwareProductValidationResultsView code too

We've responded to the comments on that file. Basically the code is written to compensate for non-standard API response from backend. The accordions are build based on different formats for input data and its a bit tricky to move it under one function as the list of conditions is different. We had discussed this with Ofir and it was agreed that we'll refactor the source in next release after backend API response is standardized by VTP.

Patch-set: 2
Reviewer: Gerrit User 5282 <5282@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0