Update patch set 3
Patch Set 3: Code-Review+2

Krystian, thank you for the quick fix.

We wanted to upgrade Swagger also for solving security vulnerabilities with old Swagger, which is now back just to stabilize SDC BE. 
Arindam, if you're able to check the option to remove old Swagger dependencies altogether, it would be great.

Patch-set: 3
Label: Code-Review=+2