Add support for HTTPS in sanity tests
If 'disableHttp' is enabled then HTTP ports are replaced with HTTPS ports.
Also added the latest trustore with AAF issuer. And because certs for SDC
components have CN set to "sdc" - environment json must have the domain
names correctly too...it will not work with plain IPs.

The changes have been tested from CSIT perspective so if something is
missing here or wrong then it is due to the fact that I was focusing on
bringing https support to the same level as are http tests and there is
also the fact that I do not know SDC component at all.

The changes are also done in a manner of previous ones - hacked on top of
the current codebase, which is in dire need of a rewrite:

- duplicated trustores and other artifacts
- duplicated scripts and/or similar script with small differences
- ui tests use old ubuntu image vs api tests on alpine
- chef codebase is very messy and I see no benefit in the current usage

Issue-ID: SDC-2564

Signed-off-by: Petr Ospalý <p.ospaly@partner.samsung.com>
Change-Id: I32e0fb7723ab8489eb09372d410c6f34aaf7cfc8