Add non-mano software information to sanity test PNF packages
Adds a sample non-mano artifact onap_pnf_software_information to
existing api sanity test PNF packages.
Resigns the 2 signed packages with the new root certificate. 2/3
packages are signed, so a new root certificate had to be generated,
as the private key of the previous one is missing.
In the docker_run script, creates a new folder to the test certificate,
different from the production folder, and bind it with the sdc-api-tests
docker.
Minor changes in docker_run script.

Change-Id: I78b28303215912df129a0e0d70afad5f3d0c2d09
Issue-ID: SDC-2641
Signed-off-by: andre.schmid <andre.schmid@est.tech>