Fix welcome screen close timeout
The welcome screen close timeout is triggering even if you close the
welcome window. After 4 seconds it forces the user to go to the home
screen even if it is already in another screen. The timeout was also
the cause for one known problem in the sanity ui test
"onapOnboardVNFflow".
Adds also a print screen to the ui tests, to check if the resource
was really clicked.
Removes invalids "throws Exception" declarations that were being
handled in the main test method "Vf.changeInstanceNameInVfTest", that
was catching also assertions errors.

Change-Id: Idff26b1ba5b199482e68cc50532fa20f4f834e34
Issue-ID: SDC-2648
Signed-off-by: andre.schmid <andre.schmid@est.tech>