Update patch set 3
Patch Set 3: Code-Review-1

I forgot to mention:
We must include a logic to check that service category is allowed. I agree hard coding it is not the best option, but we can add the allowed categories to the configuration.

Anyway we should definitely enforce that a service cannot be created under the original categories.

Patch-set: 3
Label: Code-Review=-1