Update patch set 1
Patch Set 1:

I asked our UX and product team;
We think it would be more clear to the user if the "required" field will be a new column in the inputs tab (for services only), because it's only for run time.
So we need to make sure there's no change in resources (VF or PNF) behaviour, and move the checkbox to the inputs tab instead inside add property modal.

Patch-set: 1