Configure a new Artifact Type
Allow the configuration of a new artifact type without
the need of code changes. The artifact types are
already configured in some configuration properties,
but currently the code uses an Enum to certify if
the artifact is valid.
The configuration file now is used as a source of
artifacts types instead the Enum.

Change-Id: Id0383d9fca9bce0519a4d52a4ecb3a68c8713f0f
Issue-ID: SDC-2754
Signed-off-by: andre.schmid <andre.schmid@est.tech>