Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Supporting upgrade will require lots of effort, we do have infrastructure but currently SDC support a clean install alone.
> Therefore I believe the fix is actually to revert typesToUpgrade.json file

I see, it was something introduced by the Catalog UI update. I was thinking it was something older that became broken with the update.
If you think it is better to rollback the file, I can do that in this commit. It will probably fix the bug I raised also.

Patch-set: 1