Fix logging date format to avoid parse error
When MDCSetup class in logging-filter-base tries to calculate an elapsed
time, it expects a datetime is given in ISO format.
But date format in LogFieldsMdcHandler is not compatible with ISO format
therefore 'Unable to calculate elapsed time due to error' occurs.

This change fixes date format in the class
so that the date string can be parsed by logging-filter.

Change-Id: I07c8f103f46493fe0764f904f3f00f66f35dbcf6
Issue-ID: SDC-2767
Signed-off-by: Satoshi Fujii <fujii-satoshi@jp.fujitsu.com>