Update patch set 4
Patch Set 4:

It is not likely that services exported from SDC in one operators deployment will be imported in the SDC deployment of another operator. Why then would there be a need to align between all deployments. 
We are not proposing to change or remove any of the existing core artifact types, which will continue to be aligned. We are only seeking to allow additional artifact types for which support will be desired in some deployments. 
This will allow the flexibility to include artifacts that may be required by a component in one system that is not used in other systems.
It is an acceptable and understandable limitation on this customization that export/import will only be supported across deployments with the same configuration of artifact types.

Patch-set: 4