Update patch set 2
Patch Set 2:

There are no code updates needed in SO to make use of this as SO has a plugable architecture. Anybody can write a building block that can be deployed into SO to take advantage of this (or any other) feature.
Also, SO can pass a VNF package to a VNFM through the vnfm adapter, in which case the VNFD is handled by the VNFM. The VNFM may make use of this feature

Patch-set: 2