Update patch set 1
Patch Set 1:

I thought that this patchset was applied in https://gerrit.onap.org/r/c/sdc/+/102596 commit in the patchset 3 but unfortunately that patchset was empty.
For that reason currently, the sdc-BE-init will be failing when will try to updateNormative (re-run the docker_run.sh). The reason is that onap/base_sdc-python doesn't have the yaml python package. So we can instead of catching the 'yaml.YAMLError' we can simply catch the general Exception, or update the base_sdc-python to had the yaml package:
https://gerrit.onap.org/r/gitweb?p=sdc/sdc-docker-base.git;a=blob;f=base_sdc-python/Dockerfile;h=96e4eb8a8e214536d3181f896fc8f26aabb1c8d4;hb=refs/heads/master#l22

I apologize for my mistake.

Patch-set: 1
Reviewer: Gerrit User 5335 <5335@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0