Update patch set 1
Patch Set 1:

> Thanks for fixing it;
 > Were you able to confirm 15 is enough? (I believe it is, just
 > wondering)
 > 
 > BTW please review: https://wiki.onap.org/display/DW/Commit+Messages
 > First line of commit message should be short description, no need
 > for ticket number
 > 
 > Thanks

Hi Ofir,

I have changed the commit message and removed the ticket number.

No i havent got to run through the testing just yet to ensure it works. Was just trying to get this in quickly so it can be merged and hopefully check then it is working correctly

Patch-set: 1
Label: Code-Review=0