Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Could you please look at these changes? This is my first contribution using Gerrit, and I just want to make sure I have not done anything wrong.
> 
> Also, unfortunately, resolving this issue required a quiet amount of refactoring with significant updates (I had around 30 commits initially that I have grouped in 7 to reflect the intent of each set of modifications). I hope that is alright.
> 
> Thank you in advance for your help and mentoring.

Hi, thanks for handling this ticket, I appreciate that.
Please excuse any delays with code reviews (it's a holiday for my team at at&t) I believe we'll get to it on Sunday. Of course anyone in the community is welcome to review this too in the meantime.

Ofir

Patch-set: 3
Reviewer: Gerrit User 653 <653@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>