Remove csvReportFilePath from ValidationConfigManager
This commit aims to extract the csvReportFilePath field from ValidationConfigManager in order to improve testability. This commit contains multiple minor modifications resulting from bubbling up the csvReportFilePath static field to where it is actually set. Put differently, most of the modifications result from changing the signature of the functions which formerly relied on ValidationConfigManager.getCSVReportFilePath.

Change-Id: I82c3b0fca8a0f407319e40ecfff45e89ec1d4323
Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Issue-ID: SDC-2499