Remove outputFullFilePath from ValidationConfigManager
This commit aims to extract the outputFullFilePath and outputFilePath fields from ValidationConfigManager in order to improve testability. This commit contains multiple modifications resulting from bubbling up the outputFullFilePath static field to where it is actually set. Most of the modifications result from changing the signature of the functions which formerly relied on ValidationConfigManager.getOutputFullFilePath, ValidationConfigManager.setOutputFullFilePath and ValidationConfigManager.getOutputFilePath.

Change-Id: I0a28e527cabe606de8af62b8dcbc4b4e70db98ed
Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Issue-ID: SDC-2499