Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> too many "formatting", hard to find "real" change, I'd suggest ONLY format code changed by you, and not whole file.

I initially wanted to put the formatting itself in a different commit but was told not to do it. Changing the formatting of only the code I modified would probably inconsistency in terms of coding style but if the community is good with files not respecting the code style, I'll yield.

Patch-set: 1