Update patch set 11
Patch Set 11:

Hi,

I have just pushed the last fixes and modifications and removed all the noise in order to make this change focused on adding `Try`. The remaining code will be pushed with another change once this one is validated.

Keep in mind that the goal here is to replace `Either` by a data-structure that is easier to reason about and which provides the same semantic than exception handling.

Thank you in advance for your help,

Francis.

Patch-set: 11
Work-in-progress: false