Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> VID and SO at the least, possibly more.
> If they commit to align to this change we can proceed, as long as a migration task will be delivered too for existing environments

Hi Ofir,
I was looking at the SDC tosca parser code and it looks like it converts the metadata values to String before returning to any application reading it so there would be no impact, see getValue() and getAllProperties() methods here, both return the metadata propoerty value as a String:
https://gerrit.onap.org/r/gitweb?p=sdc/sdc-tosca.git;a=blob;f=jtosca/src/main/java/org/onap/sdc/toscaparser/api/elements/Metadata.java;h=f3de49ea95ed58424160119e93c31d0c752491db;hb=ff780791b9c8b2af2aaf5d14d250b8612928b273
If Im reading this correct then there would no impact on applications in which case it could be safe to proceed with this change. Let me know your thoughts. Thanks

Patch-set: 2