Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> shouldn't this be added to the additional configuration files as well under the catalog-be src folder so that it will be picked up by the tests?

I can update the other configuration files to mimic the official one, no problem. But all tests are passing, which means they are not using this specific update. Most of the configuration files used in tests are full of not used configurations. I personally prefer to leave the test configuration files as small as the tests need.

Patch-set: 1