Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Good work removing that Either. Have you checked exception\error flow as well?

Hi,

The either is technically not removed, I have just refactored the code so it is better used, so the flow has not changed. I only leveraged the composition capabilities of Either by using `left().map`, which enables us to apply a function on the left value only if the Either in question is actually a left. Otherwise, the function is never applied.

Patch-set: 2