Refactor CsarUtil::getLatestSchemaFilesFromCassandra
This commit aims to refactor the CsarUtil::getLatestSchemaFilesFromCassandra function. It decomposes the initial function into smaller ones in order to make it more readable. On the long term, the resulting functions may be isolated in their own class, but for now, we keep these in the CsarUtil class in order to keep the refactoring small.

Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Change-Id: I3f96d76a07fd32bdbd8b59beb5409eab0b0f3aa8
Issue-ID: SDC-2812