Refactor CsarUtils::WriteArtifactDefinition
This commit aims to refactor the CsarUtils::WriteArtifactDefinition function and to use Either as recommended by the best practices. This refactoring required to break the `WriteArtifactDefinition` function in multiple smaller functions, as it was coded in a very imperative way which prevented getting the best of Either. The code is now simpler to read and to maintain. Also, a test had to be removed (CsarUtils::testWriteArtifactDefinition) as the function it covered does not exist anymore. The new functions are covered by the other tests as shown by Sonar.

Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Change-Id: I167a05d8fbc2085a3e25eeb5bfd6b30f0e80462e
Issue-ID: SDC-2812