Refactor CsarUtils::collectComponentTypeArtifacts
This commit is a minor refactoring required to refactor CsarUtils::collectComponentInstanceArtifacts. In order to keep this commit small, we only focused CsarUtils::collectComponentTypeArtifacts, and removed any output that is not reflected by the function's result type. Modifying an argument is a actually a bad practice as explained in Clean Code by Robert Martin (cf. "Output arguments should be avoided"). This commit aims to enforce this aspect.

Change-Id: Ice4d6c9a78e7706c639dd60bc272253e298bc7be
Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Issue-ID: SDC-2812