Extract checkArtifactInComponent and checkArtifactInResourceInstance from ArtifactsBusinessLogic
This commit aims to extract the checkArtifactInComponent and checkArtifactInResourceInstance functions from the ArtifactsBusinessLogic class to shrink it a bit. Tests have been added along with a some utility functions designed to generate randomized test data. The former tests were deleted as they were not doing any assertions and relied on reflection which is discouraged by Java 11.

Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Change-Id: I1b27c55d064cf4e9a40cd7b152e7ea6032a09276
Issue-ID: SDC-2961