Refactor tosca representation retrieval in CsarUtils::populateZip
This commit aims to simplify CsarUtils::populateZip by extracting how a tosca representation is fetched to another function. this enables to reduce the overall cyclomatic complexity of the function. Some additional work still needs to be done especially regarding the ToscaRepresentation and MainYamlWithDependencies classes (These happen to be really similar). This will be done in another change in the near future.

Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Change-Id: I93600d0752565de523cf2285b1fa2ba43a96334f
Issue-ID: SDC-2812