Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Did you confirm it's really in use only in HEAT only? We should double check that.
> 
> Still I have to agree with Ojas we can't tell right now where we need those annotations. VID and SO read them for sure so it must be well investigated before merged.

Hi Ofir / Ojas,
I could not find any reference for this file on the other components but I agree with you we need to test it more to make sure no regression will be introduced on other components. 

As this yml is not TOSCA complaint, it causes issues on other Tosca parses, we could at least consider to have an option to include it or not on the generated csar. For that, I can make it configurable, which will not change the actual behaviour.

Patch-set: 2