Update patch set 2
Patch Set 2:

Hi Chaya,

We have been thinking about your comment along with my co-workers and think it may be interesting to share some of these ideas with the rest of the community (if that's alright). So far we have been discussing the pros and cons of each approach (Either and exceptions) but did not really focus on the problem itself which is first about figuring out how to refactor the code in a safe and efficient way and then to make it more maintainable in the long run.

As mentioned by Ofir last Monday, the CsarUtils and ArtifactsBusinessLogic classes have a very high level of technical debt making them hard to test and maintain. This debt can be reduced by regrouping elements that are strongly related to each other, eliminating duplications, and making the code simpler, more readable and easier to reason about. However, this has to be done safely without introducing any changes to the code's behaviour.

One way to achieve this is to write more tests. However adding more tests, also means adding more code to maintain, potential bugs to fix, and can be very hard to do depending on the complexity of the code. In the case of ArtifactsBusinessLogic, we have to rely on reflection and break encapsulation, or complicated mocking strategies to cover less than 45% of the code. The reason why developers are usually reluctant to add more tests is that in the case of classes such as this one, the time and energy required outweighs the resulting benefits.

Another approach is to leverage the type system. By ensuring that a function's output is fully captured by its returned type, we can use the compiler to warn us about any potential situations we have not handled.

Null variables are a perfect example. As you know, a null-check is required whenever a function returns a null output. However, this null-check happens to be missing in many cases in SDC as shown by Sonar. This is because, from the caller’s perspective, nothing indicates a null may be produced once the function is run. We rely on the developer's discipline to ensure that any null resulting from a function is properly handled. If this is not done, the compiler won't produce any error and let the NullPointerException propagate at runtime.

A solution to this issue is to use Options. Returning an Option is a clear indication that a function may not produce any valid output given certain arguments. This forces the caller to deal with this possibility and to tell the compiler about what should be done when getting a null back.

Another example is unchecked exception throwing (like RuntimeExceptions). From the caller perspective, nothing indicates that a function may throw a runtime exception once called. This could be specified in the function's signature but in practice, the compiler won't produce any warning if this is not done. It is once again up to the author of the function to indicate this information, and up to the caller to deal with the exception. As a result, the exception may be handled very far from where it has been thrown if handled at all, and with the minimal information about why it occurred.

To sum up, the biggest challenge when refactoring a function is to figure out what it does in the first place. Currently, most of the functions in CsarUtils or ArtifactsBusinessLogic have to be inspected in order to understand what they do.  This could be improved by enhancing their signature and be explicit about what they really do using the type system.

So instead of returning nulls, these functions would return Options. Instead of throwing RuntimeExceptions, they would return a Try or an Either. Having said that, we understand your point, and we could find a middle ground. We suggest we make the distinction between the internals (private functions) and the exposed functions (public). The internals would be managed using Eithers while the public functions would throw exceptions. This would establish a clear distinction between the public API of the class and its internal domain, and ensure that exceptions are only thrown at some specific locations in the class.

Finally, we do not feel like we can efficiently refactor without creating functions as described above. This approach is based on well documented best practices, sound engineering principles, and 10 years of personal experience successfully refactoring complex codebases. Let me know your thoughts.

Thank you very much for your help and patience,

Francis

Patch-set: 2