Update patch set 2
Patch Set 2:

Hi Francis,

First, I want to thank you for the time and effort being taken to try to come up with the most appropriate and effective solution.

We've re-discussed deeply our refactoring approach and specifically error handling approach with the team including our java experts, following your comments, and I'll share our conclusions below:

- We agree that adding tests to these complicated classes isn't cost effective in many cases.
Having said that, during refactoring, usually we separate one big method into few smaller ones. In many cases the smaller methods are easier to test, and we encourage adding unit tests as much as possible.

- We agree that meaningful and well declared methods are important for refactoring.
  
- Regarding null-values, our approach is to handle null values in the edges instead of propagating them and handling them throw the whole flow. 
For instance, if we have a method that return a list "getDeploymentArtifacts()", if case deployment artifacts don't exist, the method should return an empty list instead of null.
DTO objects shouldn't be filled with null values.
In cases that this is not possible, and null value must be returned, it should be handled as close as possible (in the next 1-2 levels), in these cases Options can be used.
There's lot of work needs to be done in our code in this area, and this is the approach we push and believe in.

- Regarding error handling, you've mentioned two points to prefer Either on Exceptions:
1. "Runtime exceptions are not declared in the method signature and the developer might not think of it when using the method" - this is the developer responsibility, when she uses a method she should think about all relevant possibilities, and if there's an option to an error in the method called which can be recovered, the developer should think of it and handle it.
2. "Due to #1 - exceptions will be treated very far from where they were thrown" -as I've explained before, in most of the error cases, this is exactly the behaviour we want. Errors that fails the flow and should result in an http error shouldn't be handled all the way through the flow. We want to throw an exception where it's occurred, and catch it in the exception mapper, there, we'll log it and will convert it to the appropriate http error response.
For errors that should be handled by the application (meaning there's a way to recover from the error and continue with the flow), it's the developer responsibility to handle it as close as possible to where the error occurred.
 
In conclusion, our error handling approach is as I've explained before:
Errors that needs to be handled by the application, can be handled with Either with a correct use, or better use Try as it's biased (Vavr has a good implementation), in case it's only 1-2 levels up - otherwise - throw an exception.
For errors that stops the flow and will be converted to an http response - Either won't be acceptable, we throw an Exception that will be caught by the ExceptionMapper. This way we keep our code clean.

We highly appreciate your willing to help refactoring and will be glad you continue contribute following the guidlines above.

Thanks, 
Chaya

Patch-set: 2
Reviewer: Gerrit User 4928 <4928@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>