Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can you please try to avoid using Either here?

As mentioned in another change request, this is not the goal of the task here. I split the existing code in multiple smaller methods and tried to keep it as functionally close to the initial code as possible.
Getting rid of Either would probably make the whole code review harder than it currently already is.

Patch-set: 1