Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Can you please try to avoid using Either here?
> > 
> > Hi Idan,
> > 
> > Could you please provide reasons why to remove the Either here?
> > 
> > Thank you in advance,
> > 
> > Francis.
> 
> The goal is to reduce Either usage across the code

Hi Idan,

Could you please point the JIRA covering this issue? As far as I know, we decided to use exceptions and either in specific cases, and to refactor the usage of either so that they are used according standards. This is explained here: https://jira.onap.org/browse/SDC-2812.

Secondly, as explained by Chris, this change is not related to this task and therefore should be addressed separately.

Does this answer your question? Please reach out if there is any misunderstanding.

Thank you,

Francis.

Patch-set: 1