Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> @Ofir
> Do you think it is reasonable to assume that no other TOSCA directives will be added in the future? We cannot expect developers of new features to know this specific behavior where only one of two options is supported. I think we should either add a UI test that will test the "selecteable" and the new directive so it will fail on new future directives or take into account that we might have side effects that cause unintended backwards incompatibility

Hi All, the implementation at the moment is hardcoded and only supports setting a single Directive value "Selectable". 
What we are proposing on this change is the option to set any of those directives values (Select, Selectable, Substitute and Substitutable) from a dropdown list.
We are not changing the actual behaviour of it, but it`s clear that there are additional improvements that need to be done to enable multiples values on the directive section since it is defined as a list on the Tosca spec but as it is not the focus of this change we have proposed to have a separate user story for that.
I agree with Ilana that a new UI Test case might be required for that and it can also be addressed on a separate user story.

Patch-set: 4