Decouple TXT Report file writing and formatting logic (1/6)
This commit is the first of a series aiming to decouple the writing and formatting logic of text report files in the asdc tool. Unfortunately because of a significant amount of technical debt, a small hack had to be done so that the header of the file is not written twice. This will be removed in a future change.

Signed-off-by: Francis Toth <francis.toth@yoppworks.com>
Change-Id: I8d2a7c1353d59562d8026088824c26ae0daa36e9
Issue-ID: SDC-2499